Spectral integral and table done!
I implemented the two things we wanted last time: a value for the spectral integral from -2 to 2 km/s and a table showing the results. Both of them a created by separate functions: make_table and spectral_integral. The functions run and provide the expected results. You will find them at the bottom of the code I provide here.