L1455: 10 brightest, 5 faintest spectral bins
Turns out I actually send you yesterday all the spedtra for the bins of L1455. But I made a mistake in the labeling of the figures. so here they are again, correctly labeled.
Regardig the logical operation: if I haven't missunderstood you, I can only set the one limit of the of the interval, e.g. offset_velocity < 2. I can't give in the function call -2 < offset_velocity < 2