Code with table being plotted
In this code it has been added the function make-table2 that creates the longer table we wanted last time (see text-file). Should I keep the first 3 columns of the table? They do not change their content, so we might just put them in the title of the table. How to get the title of the table plotted in the notebook?
The issue with bin# = 0 is still not solved, thats why the strange values.