Code and table of today's task curriculum
The tasks of today are implemented: NGC1333 istead of L1455, median temperature added in the table, spectral integral multiplied by spectral-bin-width for proper scaling, no roundings of the spectral integral.
The calculated median temperature looks too big - order of x10^22. It might be that the unit is eV instead of K. I used the map 'NGC1333_Temperature.fits'. Some of the spectral integrals are negative - we might need to scale that.
And the brightness bin#=0 is still showing bad value. I haven't solved the problem there yet, there is still just 1 pixel in the bin.