correct Nchan, no more .where function, table with errors included
Here the improvements since yesterday. I uploaded the improved code (see line 189 -195 for the new implementation of Nchan and if it is correct). There is no more the .where function used. I also uploaded a plot of the spectral integrals, this time with real size error bars (no enhacement *10). And I also uploaded a new value table with the errors for every integral included.