OrionA first plots and spectral integrals
Hi, I uploaded the notebook of Orion. You will find there the spectra, the plot of the spectral integrals and also the value-table with the error included (real size error values, no enhancement). I also took account of the hyperfine structure and included it in the spectral integrals. However, the noise interval I used for the integrals of both spectra was rather short.
I also had a look on the error bars and how their size develops with the brightness bin number (0, 1, 2, ...., 14). It turns out that the errorbars are largest for the brightest bins (0 and 1), not for the faintest (14, 13, 12...). Is this still consistent with that, what you mentioned on our talk before lunch?