NH3_33 spectrum included
I added the missing spectrum. You can have a look at the plot and at the spectral integrals in the uploaded notebook. There is something that is puzzeling me: HC5N was showing for bin#=4,5,6 a clear emission line. But the integral in the margins around this feature shows no contribution in the spectral-integral-plot at the bottom of the notebook.