Spectral fit code, plots
I changed your code template a at two places, following the structure of the example codes in pyspeckit. 1) you will see that I have commented out the transformation km/s -> Hz and instead it happens automatically in the code. 2) I also defined the xarr in a different way, following this: http://pyspeckit.bitbucket.org/html/sphinx/nh3_tau.html 
For NH3_11 there is good fit, for NH3_33 I guess it is also acceptable, but for NH3_22 it is quite bad. And suprisingly the noise of the plotted spectra is much higher, than in the plot created by my function "plot_spectra", despite that the same data-array is used in both cases.