Chemical abundance, temperature, line width
Here are the plots we wanted. The chemical abundance curve behaves as you expected, more or less. The temperature, too. I provide 2 plots of the temperature, only difference is the number of bins used (data points extention towards weaker lines) for the sake of more clarity of the curve behaviour. The line width has also a for that qualitatively resembles what you scetched briefly, but quantitatively different.
And I also provide the code. The new snippet is in the dunction make_table_and_plot. Particularly, i am not sure if I properly did the log-calculations for the chemical abundance.