Plots with error bars, T_kin and bin_num=500 for line_width
Here are the corrected plots. When I had again a look at the temperature, I wondered why the slope does not go in the oposite direction (shouldn't the temperature drop with increase of the density - the star forming cores being colder than their field)?
With T_kin instead of T_exc we no longer have in the plot the sudden jump of the curve for low log(N(H2)).