Progressive bin sizes, two versions: 50*2^i and 125*2^i
Here are the plots showing what the progressive bin size provides. One of them goes in steps of 50*2^i, so 50pixels/bin, 100, 200, 400, 800, up to 1600 where I set the maximum bin size to be. And the number of bins (number of dots in the plots) are: 32bins*50pixels/bin, 16*100, 8*200, 4*400, 2*800 and all the rest bins will have the maximum bin size 1600pixels/bin.
The other plot goes in steps 125*2^i, means 8bins*125pixels/bin, 4*250, 2*500, the rest of the bins being 1600pixels/bin.
Eventually I also uploaded a plot made with the old fixed-bin-size of 500pixels/bin, so you can see a comparison.

You will see that 50*2^i provides more profile and features, on the expense of more bins and longer computational time. And I don't know how physically realistic and significant this features are.
So, what step size should I implement in the code - 50*2^i or 125*2^i?